refactor: add overflow property for NumberFormatSubmenu
