+++ src/react/loom-app/header-cell-edit/number-format-submenu/styles.css
@@ -1,0 +1,6 @@
+.dataloom-number-format-submenu {
+    width: 100%;
+    height: 240px;
+    overflow-y: scroll;
+    overflow-x: auto;
+}
